+++ lsp-ksrpc/src/commonMain/kotlin/com/monkopedia/lsp/DefaultLanguageServer.kt
@@ -22,12 +22,12 @@ open class DefaultLanguageServer : KsrpcLanguageServer {
 
     override suspend fun textDocumentImplementation(
         params: ImplementationParams
-    ): TextDocumentImplementationResult =
+    ): TextDocumentImplementationResult? =
         throw NotImplementedError("textDocumentImplementation not implemented")
 
     override suspend fun textDocumentTypeDefinition(
         params: TypeDefinitionParams
-    ): TextDocumentTypeDefinitionResult =
+    ): TextDocumentTypeDefinitionResult? =
         throw NotImplementedError("textDocumentTypeDefinition not implemented")
 
     override suspend fun textDocumentDocumentColor(
@@ -40,82 +40,83 @@ open class DefaultLanguageServer : KsrpcLanguageServer {
     ): List<ColorPresentation> =
         throw NotImplementedError("textDocumentColorPresentation not implemented")
 
-    override suspend fun textDocumentFoldingRange(params: FoldingRangeParams): List<FoldingRange> =
+    override suspend fun textDocumentFoldingRange(params: FoldingRangeParams): List<FoldingRange>? =
         throw NotImplementedError("textDocumentFoldingRange not implemented")
 
     override suspend fun textDocumentDeclaration(
         params: DeclarationParams
-    ): TextDocumentDeclarationResult =
+    ): TextDocumentDeclarationResult? =
         throw NotImplementedError("textDocumentDeclaration not implemented")
 
     override suspend fun textDocumentSelectionRange(
         params: SelectionRangeParams
-    ): List<SelectionRange> =
+    ): List<SelectionRange>? =
         throw NotImplementedError("textDocumentSelectionRange not implemented")
 
     override suspend fun textDocumentPrepareCallHierarchy(
         params: CallHierarchyPrepareParams
-    ): List<CallHierarchyItem> =
+    ): List<CallHierarchyItem>? =
         throw NotImplementedError("textDocumentPrepareCallHierarchy not implemented")
 
     override suspend fun callHierarchyIncomingCalls(
         params: CallHierarchyIncomingCallsParams
-    ): List<CallHierarchyIncomingCall> =
+    ): List<CallHierarchyIncomingCall>? =
         throw NotImplementedError("callHierarchyIncomingCalls not implemented")
 
     override suspend fun callHierarchyOutgoingCalls(
         params: CallHierarchyOutgoingCallsParams
-    ): List<CallHierarchyOutgoingCall> =
+    ): List<CallHierarchyOutgoingCall>? =
         throw NotImplementedError("callHierarchyOutgoingCalls not implemented")
 
     override suspend fun textDocumentSemanticTokensFull(
         params: SemanticTokensParams
-    ): SemanticTokens = throw NotImplementedError("textDocumentSemanticTokensFull not implemented")
+    ): SemanticTokens? = throw NotImplementedError("textDocumentSemanticTokensFull not implemented")
 
     override suspend fun textDocumentSemanticTokensFullDelta(
         params: SemanticTokensDeltaParams
-    ): TextDocumentSemanticTokensFullDeltaResult =
+    ): TextDocumentSemanticTokensFullDeltaResult? =
         throw NotImplementedError("textDocumentSemanticTokensFullDelta not implemented")
 
     override suspend fun textDocumentSemanticTokensRange(
         params: SemanticTokensRangeParams
-    ): SemanticTokens = throw NotImplementedError("textDocumentSemanticTokensRange not implemented")
+    ): SemanticTokens? =
+        throw NotImplementedError("textDocumentSemanticTokensRange not implemented")
 
     override suspend fun textDocumentLinkedEditingRange(
         params: LinkedEditingRangeParams
-    ): LinkedEditingRanges =
+    ): LinkedEditingRanges? =
         throw NotImplementedError("textDocumentLinkedEditingRange not implemented")
 
-    override suspend fun workspaceWillCreateFiles(params: CreateFilesParams): WorkspaceEdit =
+    override suspend fun workspaceWillCreateFiles(params: CreateFilesParams): WorkspaceEdit? =
         throw NotImplementedError("workspaceWillCreateFiles not implemented")
 
-    override suspend fun workspaceWillRenameFiles(params: RenameFilesParams): WorkspaceEdit =
+    override suspend fun workspaceWillRenameFiles(params: RenameFilesParams): WorkspaceEdit? =
         throw NotImplementedError("workspaceWillRenameFiles not implemented")
 
-    override suspend fun workspaceWillDeleteFiles(params: DeleteFilesParams): WorkspaceEdit =
+    override suspend fun workspaceWillDeleteFiles(params: DeleteFilesParams): WorkspaceEdit? =
         throw NotImplementedError("workspaceWillDeleteFiles not implemented")
 
-    override suspend fun textDocumentMoniker(params: MonikerParams): List<Moniker> =
+    override suspend fun textDocumentMoniker(params: MonikerParams): List<Moniker>? =
         throw NotImplementedError("textDocumentMoniker not implemented")
 
     override suspend fun textDocumentPrepareTypeHierarchy(
         params: TypeHierarchyPrepareParams
-    ): List<TypeHierarchyItem> =
+    ): List<TypeHierarchyItem>? =
         throw NotImplementedError("textDocumentPrepareTypeHierarchy not implemented")
 
     override suspend fun typeHierarchySupertypes(
         params: TypeHierarchySupertypesParams
-    ): List<TypeHierarchyItem> =
+    ): List<TypeHierarchyItem>? =
         throw NotImplementedError("typeHierarchySupertypes not implemented")
 
     override suspend fun typeHierarchySubtypes(
         params: TypeHierarchySubtypesParams
-    ): List<TypeHierarchyItem> = throw NotImplementedError("typeHierarchySubtypes not implemented")
+    ): List<TypeHierarchyItem>? = throw NotImplementedError("typeHierarchySubtypes not implemented")
 
-    override suspend fun textDocumentInlineValue(params: InlineValueParams): List<InlineValue> =
+    override suspend fun textDocumentInlineValue(params: InlineValueParams): List<InlineValue>? =
         throw NotImplementedError("textDocumentInlineValue not implemented")
 
-    override suspend fun textDocumentInlayHint(params: InlayHintParams): List<InlayHint> =
+    override suspend fun textDocumentInlayHint(params: InlayHintParams): List<InlayHint>? =
         throw NotImplementedError("textDocumentInlayHint not implemented")
 
     override suspend fun inlayHintResolve(params: InlayHint): InlayHint =
@@ -132,7 +133,7 @@ open class DefaultLanguageServer : KsrpcLanguageServer {
 
     override suspend fun textDocumentInlineCompletion(
         params: InlineCompletionParams
-    ): TextDocumentInlineCompletionResult =
+    ): TextDocumentInlineCompletionResult? =
         throw NotImplementedError("textDocumentInlineCompletion not implemented")
 
     override suspend fun initialize(params: InitializeParams): InitializeResult =
@@ -143,89 +144,89 @@ open class DefaultLanguageServer : KsrpcLanguageServer {
 
     override suspend fun textDocumentWillSaveWaitUntil(
         params: WillSaveTextDocumentParams
-    ): List<TextEdit> = throw NotImplementedError("textDocumentWillSaveWaitUntil not implemented")
+    ): List<TextEdit>? = throw NotImplementedError("textDocumentWillSaveWaitUntil not implemented")
 
     override suspend fun textDocumentCompletion(
         params: CompletionParams
-    ): TextDocumentCompletionResult =
+    ): TextDocumentCompletionResult? =
         throw NotImplementedError("textDocumentCompletion not implemented")
 
     override suspend fun completionItemResolve(params: CompletionItem): CompletionItem =
         throw NotImplementedError("completionItemResolve not implemented")
 
-    override suspend fun textDocumentHover(params: HoverParams): Hover =
+    override suspend fun textDocumentHover(params: HoverParams): Hover? =
         throw NotImplementedError("textDocumentHover not implemented")
 
-    override suspend fun textDocumentSignatureHelp(params: SignatureHelpParams): SignatureHelp =
+    override suspend fun textDocumentSignatureHelp(params: SignatureHelpParams): SignatureHelp? =
         throw NotImplementedError("textDocumentSignatureHelp not implemented")
 
     override suspend fun textDocumentDefinition(
         params: DefinitionParams
-    ): TextDocumentDefinitionResult =
+    ): TextDocumentDefinitionResult? =
         throw NotImplementedError("textDocumentDefinition not implemented")
 
-    override suspend fun textDocumentReferences(params: ReferenceParams): List<Location> =
+    override suspend fun textDocumentReferences(params: ReferenceParams): List<Location>? =
         throw NotImplementedError("textDocumentReferences not implemented")
 
     override suspend fun textDocumentDocumentHighlight(
         params: DocumentHighlightParams
-    ): List<DocumentHighlight> =
+    ): List<DocumentHighlight>? =
         throw NotImplementedError("textDocumentDocumentHighlight not implemented")
 
     override suspend fun textDocumentDocumentSymbol(
         params: DocumentSymbolParams
-    ): TextDocumentDocumentSymbolResult =
+    ): TextDocumentDocumentSymbolResult? =
         throw NotImplementedError("textDocumentDocumentSymbol not implemented")
 
     override suspend fun textDocumentCodeAction(
         params: CodeActionParams
-    ): List<TextDocumentCodeActionResult> =
+    ): List<TextDocumentCodeActionResult>? =
         throw NotImplementedError("textDocumentCodeAction not implemented")
 
     override suspend fun codeActionResolve(params: CodeAction): CodeAction =
         throw NotImplementedError("codeActionResolve not implemented")
 
-    override suspend fun workspaceSymbol(params: WorkspaceSymbolParams): JsonElement =
+    override suspend fun workspaceSymbol(params: WorkspaceSymbolParams): JsonElement? =
         throw NotImplementedError("workspaceSymbol not implemented")
 
     override suspend fun workspaceSymbolResolve(params: WorkspaceSymbol): WorkspaceSymbol =
         throw NotImplementedError("workspaceSymbolResolve not implemented")
 
-    override suspend fun textDocumentCodeLens(params: CodeLensParams): List<CodeLens> =
+    override suspend fun textDocumentCodeLens(params: CodeLensParams): List<CodeLens>? =
         throw NotImplementedError("textDocumentCodeLens not implemented")
 
     override suspend fun codeLensResolve(params: CodeLens): CodeLens =
         throw NotImplementedError("codeLensResolve not implemented")
 
-    override suspend fun textDocumentDocumentLink(params: DocumentLinkParams): List<DocumentLink> =
+    override suspend fun textDocumentDocumentLink(params: DocumentLinkParams): List<DocumentLink>? =
         throw NotImplementedError("textDocumentDocumentLink not implemented")
 
     override suspend fun documentLinkResolve(params: DocumentLink): DocumentLink =
         throw NotImplementedError("documentLinkResolve not implemented")
 
-    override suspend fun textDocumentFormatting(params: DocumentFormattingParams): List<TextEdit> =
+    override suspend fun textDocumentFormatting(params: DocumentFormattingParams): List<TextEdit>? =
         throw NotImplementedError("textDocumentFormatting not implemented")
 
     override suspend fun textDocumentRangeFormatting(
         params: DocumentRangeFormattingParams
-    ): List<TextEdit> = throw NotImplementedError("textDocumentRangeFormatting not implemented")
+    ): List<TextEdit>? = throw NotImplementedError("textDocumentRangeFormatting not implemented")
 
     override suspend fun textDocumentRangesFormatting(
         params: DocumentRangesFormattingParams
-    ): List<TextEdit> = throw NotImplementedError("textDocumentRangesFormatting not implemented")
+    ): List<TextEdit>? = throw NotImplementedError("textDocumentRangesFormatting not implemented")
 
     override suspend fun textDocumentOnTypeFormatting(
         params: DocumentOnTypeFormattingParams
-    ): List<TextEdit> = throw NotImplementedError("textDocumentOnTypeFormatting not implemented")
+    ): List<TextEdit>? = throw NotImplementedError("textDocumentOnTypeFormatting not implemented")
 
-    override suspend fun textDocumentRename(params: RenameParams): WorkspaceEdit =
+    override suspend fun textDocumentRename(params: RenameParams): WorkspaceEdit? =
         throw NotImplementedError("textDocumentRename not implemented")
 
     override suspend fun textDocumentPrepareRename(
         params: PrepareRenameParams
-    ): PrepareRenameResult = throw NotImplementedError("textDocumentPrepareRename not implemented")
+    ): PrepareRenameResult? = throw NotImplementedError("textDocumentPrepareRename not implemented")
 
-    override suspend fun workspaceExecuteCommand(params: ExecuteCommandParams): LSPAny =
+    override suspend fun workspaceExecuteCommand(params: ExecuteCommandParams): LSPAny? =
         throw NotImplementedError("workspaceExecuteCommand not implemented")
 
     override suspend fun workspaceDidChangeWorkspaceFolders(